requirements.txt
huggingface_hub
transformers
accelerate
bitsandbytes
langchain
openai
langchain_openai
langchain_community
langchain-huggingface
langchain-google-genai
python-dotenv
PyPDF2
setuptools
pandas
pinecone-client
tiktoken
chromadb
llama-cpp-python
einops
torch
google.generativeai
streamlit